Order of operations swapped.
I had to swap the order in which the added/updated/removed cursors were processed because of an apparently faulty implementation of TUIO on my end. This doesn't in practice change anything as far as the plugin's flow, but deals with weird frames where you can get an ADD and a REMOVE operations on the same session id.

I would suggest applying such change wherever is appropriate just to be a bit more forgiving about this.